+++ libraries/printHelpers/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.5.3] - 2026-07-17
+- fix #32, use printHelpersMT in printf().
+- add hexDumpLine8() and hexDumpLine16() - (experimental)
+- add examples
+- update readme.md
+- adjust build-CI
+- minor edits
 
 ## [0.5.2] - 2026-05-13
 - fix #30, handling infinity (also in printHelpersMT.h)